refactor: registered actions and removed circular imports
